+++ server/.env
@@ -1,5 +1,5 @@
 PORT = 6090
-DB_HOST = 192.168.2.13
+DB_SERVER = 192.168.2.214
 DB_USER = sa
-DB_PASSWORD = p@ssword
-DB_NAME = peb_db
+DB_PASSWORD = P@ssw0rd+IIT
+DB_NAME = PEB_DB
