+++ server/app.js
@@ -1,14 +1,10 @@
 import dotenv from 'dotenv';
 import express from 'express';
 import cors from 'cors';
-import db from './config/db.js';
-import { createConfigTimeTable } from './models/configModel.js';
-import { createPlanTimeTable, createRemoveSolidPlanTimeTable, createSummaryTable } from './models/plantimeModel.js';
-import { createTempPlanTimeTable } from './models/tempplanModel.js';
-import { createProductTable, createChemicalTable } from './models/productModel.js';
 import postRoutes from './routes/postRoute.js';
 import getRoutes from './routes/getRoute.js';
 import putRoutes from './routes/putRoute.js';
+import connectDB from './config/db.js';
 
 // กำหนดค่า Environment จากไฟล์ .env
 dotenv.config();
@@ -21,24 +17,7 @@ const PORT = process.env.PORT || 6090;
 app.use(cors());
 app.use(express.json());
 
-// connect กับฐานข้อมูล
-const connectDB = async () => {
-    try {
-        await db.getConnection();
-        console.log("✅ Database connected successfully");
-        createProductTable();
-        createChemicalTable();
-        createConfigTimeTable();
-        createPlanTimeTable();
-        createTempPlanTimeTable();
-        createRemoveSolidPlanTimeTable();
-        createSummaryTable();
-    } catch (error) {
-        console.log("❌ Error in connecting database : ", error);
-    }
-}
-
-// เรียกใช้งานฟังก์ชันเชื่อมต่อฐานข้อมูล
+// เชื่อมต่อกับฐานข้อมูล
 connectDB();
 
 // กำหนด Route สำหรับการทำงานกับ Recipe
