+++ server/controllers/tempController.js
@@ -1,121 +1,5 @@
-import db from '../config/db.js';
-
-export const addTempPlanTime = async (req, res) => {
-    const { product_name } = req.params;
-
-    try {
-        // ลบข้อมูลทั้งหมดใน temp_plan_time_table
-        await db.query(`DELETE FROM temp_plan_table`);
-
-        // รีเซ็ตค่า AUTO_INCREMENT
-        await db.query(`ALTER TABLE temp_plan_table AUTO_INCREMENT = 1`)
-
-        const [planTimes] = await db.query(`
-            SELECT pt.*
-            FROM plan_time_table pt
-            INNER JOIN product_master pm ON pt.product_id = pm.product_id
-            WHERE pm.product_name = ?
-        `, [product_name]);
-
-        if (planTimes.length === 0) {
-            return res.status(404).json({ message: '❌ No Plan Times found for this recipe' });
-        }
-
-        // INSERT ข้อมูลลงใน TempPlanTime
-        for (const plantime of planTimes) {
-            const query = `
-                INSERT INTO temp_plan_table (
-                    product_id, run_no,
-                    machine, batch_no, program_no,
-                    start_time, mixing, extruder_exit,
-                    pre_press_exit, primary_press_start, stream_in,
-                    primary_press_exit, secondary_press_1_start, temp_check_1,
-                    secondary_press_2_start, temp_check_2, cooling,
-                    secondary_press_exit, block
-                ) VALUES (
-                    ?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?
-                )
-            `;
-            const values = [
-                plantime.product_id, plantime.run_no,
-                plantime.machine, plantime.batch_no, plantime.program_no,
-                plantime.start_time, plantime.mixing, plantime.extruder_exit,
-                plantime.pre_press_exit, plantime.primary_press_start, plantime.stream_in,
-                plantime.primary_press_exit, plantime.secondary_press_1_start, plantime.temp_check_1,
-                plantime.secondary_press_2_start, plantime.temp_check_2, plantime.cooling,
-                plantime.secondary_press_exit, plantime.block
-            ]
-
-            await db.query(query, values);
-        }
-
-        return res.status(200).json({
-            message: '✅ Plan Times inserted into temp_plan_times_table successfully',
-        });
-    } catch (error) {
-        console.error('❌ ERROR:', error);
-        res.status(500).json({ message: "❌ Error in inserting Temp Plan Time" });
-    }
-}
-
-export const addTempMB = async (req, res) => {
-    const { product_name } = req.params;
-
-    try {
-        const [tempTimes] = await db.query(`
-            SELECT tpt.*
-            FROM temp_plan_table tpt
-            INNER JOIN product_master rt ON tpt.product_id = rt.product_id
-            WHERE rt.product_name = ?
-        `, [ product_name]);
-
-        if (tempTimes.length === 0) {
-            return res.status(404).json({ message: '❌ No Plan Times found for this recipe' });
-        }
-
-        // ลบข้อมูลทั้งหมดใน temp_plan_time_table
-        await db.query(`DELETE FROM temp_plan_table`);
-
-        // รีเซ็ตค่า AUTO_INCREMENT
-        await db.query(`ALTER TABLE temp_plan_table AUTO_INCREMENT = 1`)
-
-        // INSERT ข้อมูลลงใน TempPlanTime
-        for (const plantime of tempTimes) {
-            const query = `
-                INSERT INTO temp_plan_table (
-                    product_id, run_no,
-                    machine, batch_no, program_no,
-                    start_time, mixing, extruder_exit,
-                    pre_press_exit, primary_press_start, stream_in,
-                    primary_press_exit, secondary_press_1_start, temp_check_1,
-                    secondary_press_2_start, temp_check_2, cooling,
-                    secondary_press_exit, block
-                ) VALUES (
-                    ?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?
-                )
-            `;
-            const values = [
-                plantime.product_id, plantime.run_no,
-                plantime.machine, plantime.batch_no, plantime.program_no,
-                plantime.start_time, plantime.mixing, plantime.extruder_exit,
-                plantime.pre_press_exit, plantime.primary_press_start, plantime.stream_in,
-                plantime.primary_press_exit, plantime.secondary_press_1_start, plantime.temp_check_1,
-                plantime.secondary_press_2_start, plantime.temp_check_2, plantime.cooling,
-                plantime.secondary_press_exit, plantime.block
-            ]
-
-            await db.query(query, values);
-        }
-
-        return res.status(200).json({
-            message: '✅ New Tepm Plan Times add to Temp Plan Time Table success',
-            tempTimes
-        });
-    } catch (error) {
-        console.error('❌ ERROR:', error);
-        res.status(500).json({ message: "❌ Error in inserting Temp Plan Time" });
-    }
-}
+import { getPool } from "../config/db.js"; // นำเข้า getPool จาก db.js
+import sql from "mssql";
 
 // ดึงข้อมูล TempPlanTime ทั้งหมด
 export const getTempPlanTime = async (req, res) => {
