+++ server/models/configModel.js
@@ -1,32 +1,91 @@
-import db from "../config/db.js";
+import { getPool } from "../config/db.js"; // Assuming app.js exports getPool
+import sql from 'mssql';
 
-// สร้าง Process_Time_Table เพื่อเก็บข้อมูล
-export const createConfigTimeTable = async () => {
-    try {
-        await db.query(
-            `
-            CREATE TABLE IF NOT EXISTS config_time (
-                config_id INT PRIMARY KEY AUTO_INCREMENT,
-                config_group VARCHAR(255) NOT NULL,
-                mixing_time INT NOT NULL,
-                extruder_exit_time INT NOT NULL,
-                pre_press_exit_time INT NOT NULL,
-                primary_press_start INT NOT NULL,
-                stream_in INT NOT NULL,
-                primary_press_exit INT NOT NULL,
-                secondary_press_1_start INT NOT NULL,
-                temp_check_1 INT NOT NULL,
-                secondary_press_2_start INT NOT NULL,
-                temp_check_2 INT NOT NULL,
-                cooling_time INT NOT NULL,
-                secondary_press_exit INT NOT NULL,
-                adj_next_start INT NOT NULL,
-                solid_block INT,
-                remove_workpiece INT
-            )
-            `
+// บันทึก Config time ลงในฐานข้อมูล
+export const addConfig = async (req, res) => {
+    const {
+        config_group,
+        mixing_time,
+        extruder_exit_time,
+        pre_press_exit_time,
+        primary_press_start,
+        stream_in,
+        primary_press_exit,
+        secondary_press_1_start,
+        temp_check_1,
+        secondary_press_2_start,
+        temp_check_2,
+        cooling_time,
+        secondary_press_exit,
+        adj_next_start,
+        solid_block,
+        remove_workpiece,
+    } = req.body;
+
+    const query = `
+        INSERT INTO config_time (
+            config_group,
+            mixing_time,
+            extruder_exit_time,
+            pre_press_exit_time,
+            primary_press_start,
+            stream_in,
+            primary_press_exit,
+            secondary_press_1_start,
+            temp_check_1,
+            secondary_press_2_start,
+            temp_check_2,
+            cooling_time,
+            secondary_press_exit,
+            adj_next_start,
+            solid_block,
+            remove_workpiece
+        ) VALUES (
+            @config_group,
+            @mixing_time,
+            @extruder_exit_time,
+            @pre_press_exit_time,
+            @primary_press_start,
+            @stream_in,
+            @primary_press_exit,
+            @secondary_press_1_start,
+            @temp_check_1,
+            @secondary_press_2_start,
+            @temp_check_2,
+            @cooling_time,
+            @secondary_press_exit,
+            @adj_next_start,
+            @solid_block,
+            @remove_workpiece
         )
+    `;
+
+    try {
+        const pool = await getPool();
+        const request = pool.request();
+
+        // กำหนดค่า Parameters
+        request.input('config_group', sql.VarChar, config_group);
+        request.input('mixing_time', sql.Int, mixing_time);
+        request.input('extruder_exit_time', sql.Int, extruder_exit_time);
+        request.input('pre_press_exit_time', sql.Int, pre_press_exit_time);
+        request.input('primary_press_start', sql.Int, primary_press_start);
+        request.input('stream_in', sql.Int, stream_in);
+        request.input('primary_press_exit', sql.Int, primary_press_exit);
+        request.input('secondary_press_1_start', sql.Int, secondary_press_1_start);
+        request.input('temp_check_1', sql.Int, temp_check_1);
+        request.input('secondary_press_2_start', sql.Int, secondary_press_2_start);
+        request.input('temp_check_2', sql.Int, temp_check_2);
+        request.input('cooling_time', sql.Int, cooling_time);
+        request.input('secondary_press_exit', sql.Int, secondary_press_exit);
+        request.input('adj_next_start', sql.Int, adj_next_start);
+        request.input('solid_block', sql.Int, solid_block);
+        request.input('remove_workpiece', sql.Int, remove_workpiece);
+
+        await request.query(query);
+        res.status(201).send("Config added successfully");
     } catch (error) {
-        console.log("❌ Error in creating table 'config_table' : ", error);
+        res.status(500).send("Error in adding config");
+        console.log("❌ Error in adding config : ", error);
     }
-}
+};
