+++ server/package.json
@@ -15,6 +15,7 @@
     "dotenv": "^16.4.7",
     "express": "^4.21.2",
     "jsonwebtoken": "^9.0.2",
+    "mssql": "^11.0.1",
     "mysql2": "^3.13.0",
     "nodemon": "^3.1.9"
   }
